watchtower/wtserver: move from watchtower/server
