+++ watchtower/wtserver/interface.go
@@ -1,4 +1,4 @@
-package server
+package wtserver
 
 import (
 	"io"
