+++ watchtower/wtserver/server.go
@@ -1,4 +1,4 @@
-package server
+package wtserver
 
 import (
 	"bytes"
